requirements.txt
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
dlib==19.24.2
face-recognition==1.3.0
face-recognition-models==0.3.0
idna==3.4
numpy==1.26.1
opencv-python==4.8.1.78
pandas==2.1.2
Pillow==10.1.0
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
tzdata==2023.3
urllib3==2.0.7
